@@ -1,17 +1,12 @@
+var RecentCounter = function () {};
 
-var RecentCounter = function() {
-    
-};
-
-/** 
+/**
  * @param {number} t
  * @return {number}
  */
-RecentCounter.prototype.ping = function(t) {
-    
-};
+RecentCounter.prototype.ping = function (t) {};
 
-/** 
+/**
  * Your RecentCounter object will be instantiated and called as such:
  * var obj = new RecentCounter()
  * var param_1 = obj.ping(t)
