@@ -1,3 +1,3 @@
 function canFormArray(arr: number[], pieces: number[][]): boolean {
-    return false;
-};
+  return false;
+}
