@@ -2,8 +2,8 @@
  * @param {number[]} nums
  * @return {number}
  */
-var findShortestSubArray = function(nums) {
-    return 0;
+var findShortestSubArray = function (nums) {
+  return 0;
 };
 
 module.exports = { findShortestSubArray };
