@@ -1,3 +1,3 @@
 function maxSubArray(nums: number[]): number {
-return 0;
-};
+  return 0;
+}
