@@ -3,8 +3,8 @@
  * @param {number} target
  * @return {number[]}
  */
-var twoSum = function(nums, target) {
-    'use strict';
-    return [];
+var twoSum = function (nums, target) {
+  "use strict";
+  return [];
 };
-module.exports = { twoSum }
+module.exports = { twoSum };
