@@ -2,8 +2,8 @@
  * @param {number[]} nums
  * @return {number}
  */
-var lengthOfLIS = function(nums) {
-        return 0;
+var lengthOfLIS = function (nums) {
+  return 0;
 };
 
-module.exports = { lengthOfLIS }
+module.exports = { lengthOfLIS };
